refactoring few bits out of export
- created common module for tools ;
- created exception module to regroup exceptions
