+++ eagle_automation/exceptions.py
@@ -1,0 +1,12 @@
+#!/usr/bin/env python
+
+
+class FileNotFoundError(IOError):
+    pass
+
+
+class BadExtension(Exception):
+    pass
+
+
+
